Ship a simple JSON patch document to test with
We are going to experiment with JSON patch in order to maintain various
fixes to the original data as patches. We could maintain them in 2 other
ways, namely in the code or directly on the data. The former is bound to
become unyieldy, the latter diverges us enough from the original data to
make our data their "own" thing, which is not what I currently want.